feat: updated metadata files, remove google map
